DRC-2011: Implemented offset calculation for ankles. Started on upper body offset calculation. does not fixbamboo

+++ Acsell/src/us/ihmc/steppr/hardware/state/slowSensors/IMUGyroSensor.java
@@ -15,7 +15,7 @@ public class IMUGyroSensor implements StepprSlowSensor
    @Override
    public void update(int value)
    {
-      imuGyro.set(value);
+      imuGyro.set(((short)value));
    }
 
 }
